requirements.txt
# Madrid Food Delivery Streaming Simulation - Dependencies
# Python 3.9+

fastavro>=1.9.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
azure-eventhub>=5.11.0
